+++ tests/integration/helm_opentelemetry_operator_enabled_test.go
@@ -10,17 +10,16 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/client-go/kubernetes/scheme"
-	"sigs.k8s.io/e2e-framework/klient/k8s"
 	"sigs.k8s.io/e2e-framework/klient/k8s/resources"
 	"sigs.k8s.io/e2e-framework/klient/wait"
 	"sigs.k8s.io/e2e-framework/klient/wait/conditions"
 	"sigs.k8s.io/e2e-framework/pkg/envconf"
 	"sigs.k8s.io/e2e-framework/pkg/features"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/SumoLogic/sumologic-kubernetes-collection/tests/integration/internal/ctxopts"
+	"github.com/SumoLogic/sumologic-kubernetes-collection/tests/integration/internal/stepfuncs"
 )
 
 func Test_Helm_OpenTelemetry_Operator_Enabled(t *testing.T) {
@@ -36,67 +35,49 @@ func Test_Helm_OpenTelemetry_Operator_Enabled(t *testing.T) {
 	}
 
 	featTraces := features.New("traces").
-		// TODO: Rewrite into similar step func as WaitUntilStatefulSetIsReady but for deployments
-		Assess("traces-sampler deployment is ready", func(ctx context.Context, t *testing.T, envConf *envconf.Config) context.Context {
-			res := envConf.Client().Resources(ctxopts.Namespace(ctx))
-			releaseName := ctxopts.HelmRelease(ctx)
-			labelSelector := fmt.Sprintf("app=%s-sumologic-traces-sampler", releaseName)
-			ds := appsv1.DeploymentList{}
-
-			require.NoError(t,
-				wait.For(
-					conditions.New(res).
-						ResourceListN(&ds, 1,
-							resources.WithLabelSelector(labelSelector),
-						),
-					wait.WithTimeout(waitDuration),
-					wait.WithInterval(tickDuration),
+		Assess("traces-sampler deployment is ready",
+			stepfuncs.WaitUntilDeploymentIsReady(
+				waitDuration,
+				tickDuration,
+				stepfuncs.WithNameF(
+					stepfuncs.ReleaseFormatter("%s-sumologic-traces-sampler"),
 				),
-			)
-			require.NoError(t,
-				wait.For(
-					conditions.New(res).
-						DeploymentConditionMatch(&ds.Items[0], appsv1.DeploymentAvailable, corev1.ConditionTrue),
-					wait.WithTimeout(waitDuration),
-					wait.WithInterval(tickDuration),
+				stepfuncs.WithLabelsF(stepfuncs.LabelFormatterKV{
+					K: "app",
+					V: stepfuncs.ReleaseFormatter("%s-sumologic-traces-sampler"),
+				},
 				),
-			)
-			return ctx
-		}).
-		Assess("otelcol-instrumentation statefulset is ready", func(ctx context.Context, t *testing.T, envConf *envconf.Config) context.Context {
-			res := envConf.Client().Resources(ctxopts.Namespace(ctx))
-			nl := corev1.NodeList{}
-			if !assert.NoError(t, res.List(ctx, &nl)) {
-				return ctx
-			}
-
-			releaseName := ctxopts.HelmRelease(ctx)
-			labelSelector := fmt.Sprintf("app=%s-sumologic-otelcol-instrumentation", releaseName)
-			ds := appsv1.StatefulSetList{}
-
-			require.NoError(t,
-				wait.For(
-					conditions.New(res).
-						ResourceListN(&ds, 1,
-							resources.WithLabelSelector(labelSelector),
-						),
-					wait.WithTimeout(waitDuration),
-					wait.WithInterval(tickDuration),
+			),
+		).
+		Assess("otelcol-instrumentation statefulset is ready",
+			stepfuncs.WaitUntilStatefulSetIsReady(
+				waitDuration,
+				tickDuration,
+				stepfuncs.WithNameF(
+					stepfuncs.ReleaseFormatter("%s-sumologic-otelcol-instrumentation"),
 				),
-			)
-			require.NoError(t,
-				wait.For(
-					conditions.New(res).
-						ResourceMatch(&ds.Items[0], func(object k8s.Object) bool {
-							s := object.(*appsv1.StatefulSet)
-							return s.Status.Replicas == s.Status.ReadyReplicas
-						}),
-					wait.WithTimeout(waitDuration),
-					wait.WithInterval(tickDuration),
+				stepfuncs.WithLabelsF(
+					stepfuncs.LabelFormatterKV{
+						K: "app",
+						V: stepfuncs.ReleaseFormatter("%s-sumologic-otelcol-instrumentation"),
+					},
 				),
-			)
-			return ctx
-		}).
+			),
+		).
+		Assess("traces-gateway deployment is ready",
+			stepfuncs.WaitUntilDeploymentIsReady(
+				waitDuration,
+				tickDuration,
+				stepfuncs.WithNameF(
+					stepfuncs.ReleaseFormatter("%s-sumologic-traces-gateway"),
+				),
+				stepfuncs.WithLabelsF(stepfuncs.LabelFormatterKV{
+					K: "app",
+					V: stepfuncs.ReleaseFormatter("%s-sumologic-traces-gateway"),
+				},
+				),
+			),
+		).
 		Feature()
 
 	featOpenTelemetryOperator := features.New("opentelemetry-operator").