+++ tests/integration/internal/stepfuncs/assess_funcs.go
@@ -336,3 +336,54 @@ func WaitUntilDaemonSetIsReady(
 		return ctx
 	}
 }
+
+// WaitUntilDeploymentIsReady waits for a specified duration and checks with the
+// specified tick interval whether the deployment (as described by the provided options)
+// is ready.
+//
+// Readiness for a deployment is defined as having Status.NumberUnavailable == 0.
+func WaitUntilDeploymentIsReady(
+	waitDuration time.Duration,
+	tickDuration time.Duration,
+	opts ...Option,
+) features.Func {
+	return func(ctx context.Context, t *testing.T, envConf *envconf.Config) context.Context {
+		deps := appsv1.Deployment{
+			ObjectMeta: v1.ObjectMeta{
+				Namespace: ctxopts.Namespace(ctx),
+			},
+		}
+
+		listOpts := []resources.ListOption{}
+		for _, opt := range opts {
+			opt.Apply(ctx, &deps)
+			listOpts = append(listOpts, opt.GetListOption(ctx))
+		}
+
+		res := envConf.Client().Resources(ctxopts.Namespace(ctx))
+		cond := conditions.
+			New(res).
+			ResourceListMatchN(&appsv1.DeploymentList{Items: []appsv1.Deployment{deps}},
+				1,
+				func(obj k8s.Object) bool {
+					dep := obj.(*appsv1.Deployment)
+					log.V(5).InfoS("Deployment", "status", dep.Status)
+					if dep.Status.UnavailableReplicas != 0 {
+						log.V(0).Infof("Deployment %q not yet fully ready", dep.Name)
+						return false
+					}
+					return true
+				},
+				listOpts...,
+			)
+
+		require.NoError(t,
+			wait.For(cond,
+				wait.WithTimeout(waitDuration),
+				wait.WithInterval(tickDuration),
+			),
+		)
+
+		return ctx
+	}
+}